More PR fixes
