+++ samples/MvcSample.Web/Startup.cs
@@ -40,21 +40,6 @@ namespace MvcSample.Web
 
                 app.UseServices(services =>
                 {
-                    services.ConfigureAuthorization(auth =>
-                    {
-                        auth.AddPolicy(
-                            "CanViewPage", 
-                            new AuthorizationPolicyBuilder()
-                                .RequiresClaim("Permission", "CanViewPage", "CanViewAnything").Build());
-                        auth.AddPolicy(
-                            "CanViewAnything", 
-                            new AuthorizationPolicyBuilder()
-                                .RequiresClaim("Permission", "CanViewAnything").Build());
-                        // This policy basically requires that the auth type is present
-                        var basicPolicy = new AuthorizationPolicyBuilder("Basic").RequiresClaim(ClaimTypes.NameIdentifier);
-                        auth.AddPolicy("RequireBasic", basicPolicy.Build());
-                    });
-
                     services.AddMvc();
                     services.AddSingleton<PassThroughAttribute>();
                     services.AddSingleton<UserNameService>();
@@ -113,22 +98,6 @@ namespace MvcSample.Web
                         options.AddXmlDataContractSerializerFormatter();
                         options.Filters.Add(new FormatFilterAttribute());
                     });
-                    services.ConfigureAuthorization(auth =>
-                    {
-                        auth.AddPolicy(
-                            "CanViewPage",
-                            new AuthorizationPolicyBuilder()
-                                .RequiresClaim("Permission", "CanViewPage", "CanViewAnything").Build());
-                        auth.AddPolicy(
-                            "CanViewAnything",
-                            new AuthorizationPolicyBuilder()
-                                .RequiresClaim("Permission", "CanViewAnything").Build());
-
-                        // This policy basically requires that the auth type is present
-                        var basicPolicy = new AuthorizationPolicyBuilder("Basic").RequiresClaim(ClaimTypes.NameIdentifier);
-                        auth.AddPolicy("RequireBasic", basicPolicy.Build());
-                    });
-
                 });
             }
