+++ test/Microsoft.AspNet.Mvc.Core.Test/ControllerActionDescriptorProviderTests.cs
@@ -1363,8 +1363,9 @@ namespace Microsoft.AspNet.Mvc.Test
             IEnumerable<IFilter> filters = null)
         {
             var controllerTypeProvider = new FixedSetControllerTypeProvider(new[] { controllerTypeInfo });
-            var controllerModelBuilder = new DefaultControllerModelBuilder(new DefaultActionModelBuilder(),
-                                                                           NullLoggerFactory.Instance);
+            var controllerModelBuilder = new DefaultControllerModelBuilder(new DefaultActionModelBuilder(null),
+                                                                           NullLoggerFactory.Instance,
+                                                                           null);
 
             var provider = new ControllerActionDescriptorProvider(
                 controllerTypeProvider,
@@ -1380,8 +1381,9 @@ namespace Microsoft.AspNet.Mvc.Test
             params TypeInfo[] controllerTypeInfo)
         {
             var controllerTypeProvider = new FixedSetControllerTypeProvider(controllerTypeInfo);
-            var controllerModelBuilder = new DefaultControllerModelBuilder(new DefaultActionModelBuilder(),
-                                                                           NullLoggerFactory.Instance);
+            var controllerModelBuilder = new DefaultControllerModelBuilder(new DefaultActionModelBuilder(null),
+                                                                           NullLoggerFactory.Instance,
+                                                                           null);
 
             var provider = new ControllerActionDescriptorProvider(
                 controllerTypeProvider,
@@ -1398,8 +1400,9 @@ namespace Microsoft.AspNet.Mvc.Test
             IApplicationModelConvention convention)
         {
             var controllerTypeProvider = new FixedSetControllerTypeProvider(new[] { type });
-            var modelBuilder = new DefaultControllerModelBuilder(new DefaultActionModelBuilder(),
-                                                                 NullLoggerFactory.Instance);
+            var modelBuilder = new DefaultControllerModelBuilder(new DefaultActionModelBuilder(null),
+                                                                 NullLoggerFactory.Instance,
+                                                                 null);
 
             var options = new MockMvcOptionsAccessor();
             options.Options.Conventions.Add(convention);
@@ -1415,8 +1418,9 @@ namespace Microsoft.AspNet.Mvc.Test
         private IEnumerable<ActionDescriptor> GetDescriptors(params TypeInfo[] controllerTypeInfos)
         {
             var controllerTypeProvider = new FixedSetControllerTypeProvider(controllerTypeInfos);
-            var modelBuilder = new DefaultControllerModelBuilder(new DefaultActionModelBuilder(),
-                                                                 NullLoggerFactory.Instance);
+            var modelBuilder = new DefaultControllerModelBuilder(new DefaultActionModelBuilder(null),
+                                                                 NullLoggerFactory.Instance,
+                                                                 null);
 
             var provider = new ControllerActionDescriptorProvider(
                 controllerTypeProvider,
