+++ test/Microsoft.AspNet.Mvc.FunctionalTests/FiltersTest.cs
@@ -154,7 +154,7 @@ namespace Microsoft.AspNet.Mvc.FunctionalTests
                 "http://localhost/AuthorizeUser/Impossible");
 
             // Assert
-            Assert.Equal(HttpStatusCode.Forbidden, response.StatusCode);
+            Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
         }
 
         [Fact]
