+++ test/WebSites/FiltersWebSite/ManagerHandler.cs
@@ -1,0 +1,21 @@
+﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Security.Claims;
+using System.Threading.Tasks;
+using Microsoft.AspNet.Security;
+using Microsoft.Framework.DependencyInjection;
+
+namespace FiltersWebSite
+{
+    public class ManagerHandler : AuthorizationHandler<OperationAuthorizationRequirement>
+    {
+        public override void Handle(Microsoft.AspNet.Security.AuthorizationContext context, OperationAuthorizationRequirement requirement)
+        {
+            if (context.User.HasClaim("Manager", "yes"))
+            {
+                context.Succeed(requirement);
+            }
+        }
+    }
+}
